Create license file for Triangle.NET
Trying to make sure do what I can to preserve the history of Triangle.NET, which is certainly not my own creation but didn't have the typical accreditation/licensing at the time I added it to this project.  Adding this license and attribution should help ensure that proper credit is given where it's due.